+++ lib/profiles.go
@@ -59,33 +59,29 @@ func CreateProfileSettings(flags *pflag.FlagSet, rootProfileName *string, defaul
 
     flagConfigProvider := newFlagConfigProvider(flags, false)
 
-    profileSettings.ProfileName = evaluateString(labelProfile,
-        flagConfigProvider(FlagSetProfile),
-        interactiveStringValue(profilePrompt(rootProfileName, labelProfile), defaultSettings.ProfileName, nil))
-
     profileSettings.OktaUser = evaluateString(labelUser,
         flagConfigProvider(FlagSetUser),
-        interactiveStringValue(profilePrompt(profileSettings.ProfileName, labelUser), defaultSettings.OktaUser, nil))
+        interactiveStringValue(profilePrompt(rootProfileName, labelUser), defaultSettings.OktaUser, nil))
 
     profileSettings.OktaURL = evaluateString(labelOktaURL,
         flagConfigProvider(FlagSetOktaURL),
-        interactiveStringValue(profilePrompt(profileSettings.ProfileName, labelOktaURL), defaultSettings.OktaURL, validateURL))
+        interactiveStringValue(profilePrompt(rootProfileName, labelOktaURL), defaultSettings.OktaURL, validateURL))
 
     profileSettings.OktaAppURL = evaluateString(labelOktaAppURL,
         flagConfigProvider(FlagSetOktaAppURL),
-        interactiveStringValue(profilePrompt(profileSettings.ProfileName, labelOktaAppURL), defaultSettings.OktaAppURL, validateURL))
+        interactiveStringValue(profilePrompt(rootProfileName, labelOktaAppURL), defaultSettings.OktaAppURL, validateURL))
 
     profileSettings.OktaAuthMethod = evaluateString(labelAuthMethod,
         flagConfigProvider(FlagSetAuthMethod),
-        interactiveMenu(profilePrompt(profileSettings.ProfileName, labelAuthMethod), authMethods, defaultSettings.OktaAuthMethod))
+        interactiveMenu(profilePrompt(rootProfileName, labelAuthMethod), authMethods, defaultSettings.OktaAuthMethod))
 
     profileSettings.AwsRole = evaluateString(labelRole,
         flagConfigProvider(FlagSetRole),
-        interactiveStringValue(profilePrompt(profileSettings.ProfileName, labelRole), defaultSettings.AwsRole, nil))
+        interactiveStringValue(profilePrompt(rootProfileName, labelRole), defaultSettings.AwsRole, nil))
 
     profileSettings.AwsSessionTime = evaluateInt(labelSessionTime,
         flagConfigProvider(FlagSetSessionTime),
-        interactiveIntValue(profilePrompt(profileSettings.ProfileName, labelSessionTime), defaultSettings.AwsSessionTime))
+        interactiveIntValue(profilePrompt(rootProfileName, labelSessionTime), defaultSettings.AwsSessionTime))
 
     if err := validate.Struct(profileSettings); err != nil {
         return nil, err